+++ src/models/Settings.php
@@ -6,7 +6,6 @@ use craft\base\Model;
 
 class Settings extends Model
 {
-    public $colours = [];
     public $colors = [];
     public $palettes = [];
 }